requirements.txt
torch>=1.9.0
torchvision
regex
ftfy
tqdm
huggingface_hub
safetensors
timm
scikit-learn
braceexpand
pandas
webdataset
pyrootutils
pycm
matplotlib
mlxtend
transformers
opencv-python
pillow_heif
git+https://github.com/openai/CLIP.git
accelerate
jaxtyping
pydantic
einops
strenum
eval_type_backport